Refactor FileLibraryDialog to break out browser to separate component
