+++ src/features/files/types.ts
@@ -1,0 +1,5 @@
+export type TypeOption = 'image';
+
+export const TYPE_OPTIONS: Record<TypeOption, string[]> = {
+  image: ['image/png', 'image/jpeg'],
+};
